refactored examples
.
