+++ packages/examples/src/demo/blocks/index.tsx
@@ -1,0 +1,9 @@
+export * from "./actions";
+export * from "./benchmark";
+export * from "./blocksSwitcher";
+export * from "./configChange";
+export * from "./initFiles";
+export * from "./input";
+export * from "./output";
+export * from "./skins";
+export * from "./validation";
