+++ packages/examples/src/demo/config/index.tsx
@@ -18,12 +18,9 @@ import moment from "moment";
 import ru_RU from "antd/es/locale/ru_RU";
 import { ruRU } from "@material-ui/core/locale";
 import { ruRU as muiRuRU } from "@mui/material/locale";
+import { AntdWidgets } from "@react-awesome-query-builder/antd";
+import { skinToConfig } from "../../skins";
 
-import { AntdConfig, AntdWidgets } from "@react-awesome-query-builder/antd";
-import { MuiConfig } from "@react-awesome-query-builder/mui";
-import { MaterialConfig } from "@react-awesome-query-builder/material";
-import { BootstrapConfig } from "@react-awesome-query-builder/bootstrap";
-import { FluentUIConfig } from "@react-awesome-query-builder/fluent";
 const {
   FieldSelect,
   FieldDropdown,
@@ -32,14 +29,6 @@ const {
 } = AntdWidgets;
 const { simulateAsyncFetch } = Utils.Autocomplete;
 
-const skinToConfig: Record<string, Config> = {
-  vanilla: BasicConfig,
-  antd: AntdConfig,
-  material: MaterialConfig,
-  mui: MuiConfig,
-  bootstrap: BootstrapConfig,
-  fluent: FluentUIConfig
-};
 
 export default (skin: string) => {
   const InitialConfig = skinToConfig[skin] as BasicConfig;
@@ -281,7 +270,7 @@ export default (skin: string) => {
     },
     fieldSources: ["field", "func"],
     keepInputOnChangeFieldSrc: true,
-    reverseOperatorsForNot: false,
+    reverseOperatorsForNot: true,
     canShortMongoQuery: true,
     // canReorder: true,
     // canRegroup: true,
