+++ packages/examples/src/demo/init_data/logic/simple.js
@@ -1,0 +1,13 @@
+export default 
+{
+  "and": [
+    {
+      "==": [
+        {
+          "var": "slider"
+        },
+        35
+      ]
+    },
+  ]
+};
