+++ packages/examples/src/demo/init_data/logic/with_not_in_some.js
@@ -1,0 +1,29 @@
+export default 
+{
+  "and": [
+    { "some": [
+      { "var": "cars" },
+      { "in": [ { "var": "vendor" }, [ "Ford", "Toyota" ] ] }
+    ] },
+    { "all": [
+      { "var": "cars" },
+      { "!": { "in": [ { "var": "vendor" }, [ "Ford", "Toyota" ] ] } }
+    ] },
+    { "all": [
+      { "var": "cars" },
+      { "!": { "and": [ { "!": { "in": [ { "var": "vendor" }, [ "Ford", "Toyota" ] ] } } ] } }
+    ] },
+    { "some": [
+      { "var": "cars" },
+      { "!": { "and": [ { "!": { "in": [ { "var": "vendor" }, [ "Ford", "Toyota" ] ] } } ] } }
+    ] },
+    { "some": [
+      { "var": "cars" },
+      { "!": { "!": { "in": [ { "var": "vendor" }, [ "Ford", "Toyota" ] ] } } }
+    ] },
+    { "some": [
+      { "var": "cars" },
+      { "!": { "!": { "!": { "in": [ { "var": "vendor" }, [ "Ford", "Toyota" ] ] } } } }
+    ] }
+  ]
+};
