+++ packages/examples/src/demo_switch/config.tsx
