+++ packages/examples/src/skins/mui.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect } from "react";
-import { LazyStyleModule } from "../utils";
+import { LazyStyleModule } from "./utils";
 // @ts-ignore
-import styles from "@react-awesome-query-builder/fluent/css/styles.scss";
+import styles from "@react-awesome-query-builder/mui/css/styles.scss";
 
-const ImportFluent: React.FC = () => {
+const ImportMui: React.FC = () => {
   useEffect(() => {
     (styles as LazyStyleModule).use();
     return () => {
@@ -13,4 +13,4 @@ const ImportFluent: React.FC = () => {
   return null;
 };
 
-export default ImportFluent;
+export default ImportMui;
